pyproject.toml
[tool.poetry]
name = "url-shortener"
version = "1.0.0"
description = ""
authors = [
  "João Paulo Ramos Nogueira <joao.rnogueira@pagar.me>",
]

[tool.poetry.dependencies]
python = "~3.10"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
pydantic = "^1.10.2"
google-cloud-firestore = "^2.7.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
flake8 = "^4.0.1"
pytest-asyncio = "^0.19.0"
isort = "^5.10.1"
blue = "^0.9.1"
pre-commit = "^2.20.0"
httpx = "^0.23.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 120
skip = [".gitignore", ".env", "Makefile", "Dockerfile", ".flake8"]
extend_skip = [".tf", ".md", ".json", ".yaml", ".ini", ".lock", ".local", ".toml"]
lines_after_imports = 2

[tool.blue]
line-length = 120
